archive/flask_server/requirements.txt
cryptography==0.7.2
cssmin==0.2.0
Flask==0.10.1
Flask-Assets==0.10
Flask-SQLAlchemy==2.0
Flask-WebTest==0.0.7
Flask-WTF==0.11
grequests==0.2.0
jsmin==2.1.0
phonenumbers==7.0.2
requests==2.5.1
voluptuous==0.8.7
WTForms-Components==0.9.7
